+++ app/src/main/java/com/tnco/runar/data/local/DataDB.kt
@@ -14,6 +14,7 @@ import com.tnco.runar.model.UserLayoutModel
 )
 
 abstract class DataDB : RoomDatabase() {
+
     abstract fun dataDAO(): DataDAO
 
     companion object {
