--- app/src/main/java/com/tnco/runar/data/remote/En.kt
@@ -1,6 +1,0 @@
-package com.tnco.runar.data.remote
-
-data class En(
-    val description: String? = null,
-    val title: String? = null
-)
