--- app/src/main/java/com/tnco/runar/data/remote/Ru.kt
@@ -1,6 +1,0 @@
-package com.tnco.runar.data.remote
-
-data class Ru(
-    val description: String? = null,
-    val title: String? = null
-)
