--- app/src/main/java/com/tnco/runar/data/remote/RunesResponse.kt
@@ -1,8 +1,0 @@
-package com.tnco.runar.data.remote
-
-data class RunesResponse(
-    val en: En? = null,
-    val id: Int? = null,
-    val imgUrl: String? = null,
-    val ru: Ru? = null
-)
